chore: add script for profiling
This commit adds a script to generate profiling reports for the
kepler. The script assumes that Kepler is build and running in the
background. It runs the pprof tool to capture CPU and memory profiles
and then downloads the necessary files from the pprof http server to
view them locally in the browser.

Also a GH action is added which builds and deploy Kepler using fake cpu
power meter(VM) and then captures the profiling artifacts.
Once the profiling artifacts are ready, it uploads them and adds a
comment to the PR for the users to download the profiling artifacts from
the CI run

Signed-off-by: vprashar2929 <vibhu.sharma2929@gmail.com>